Fix broken bundles when dependency is not present.
As of now, leaving *dev* dependencies breaks the bootstrapping
because it tries to load non existent classes.

With this change, the class is not explicit and it's not autoloaded
when searching for bundles, so if your environment setting
does not need a bundle, it doesn't force you to require the
dependency.

Also, in case a bundle is misconfigured between the composer.json
and the required environment, it checks and throws a RuntimeException
if class is needed but still not found.